creating pointListController and adjust to the work for it
